split propagator versions into separate files
